[OpenViBE Module] Fix some compilation warnings in build of 64 bits in openvibe module csv, socket, xml and module tests. * Replace some unsigned int with "size_t" in several local method * Change signature of CString length method to return a size_t
Tracker reference: CORE-1270

+++ modules/csv/src/ovCCSV.hpp
@@ -332,7 +332,7 @@ namespace OpenViBE
 			std::vector<double> m_FrequencyAbscissa;
 
 			uint32_t m_SamplingRate;
-			uint32_t m_ColumnCount;
+			size_t m_ColumnCount;
 
 			bool m_HasInputType;
 			bool m_IsFirstLineWritten;
