+++ modules/socket/src/IConnectionServer.cpp
@@ -96,7 +96,7 @@ namespace Socket
 			{
 				return NULL;
 			}
-			return new TConnection<IConnection>(l_i32ClientSocket);
+			return new TConnection<IConnection>(static_cast<int32>(l_i32ClientSocket));
 		}
 
 		bool getSocketPort(uint32_t& port)
